Revise WallFollow.ino
Compeletely re-do the Wall Following algorithm to use PID for steady
state and right turns and a special if statement and turnLeft function
for hard left turns. The PID constants are more or less optimal, we could
speed this up a fair bit but it's probably not worth the effort.